# openapi.json
{
  "components": {
    "schemas": {
      "APIResponse": {
        "description": "Top-level envelope returned by every endpoint (untyped ``data``).",
        "properties": {
          "data": {
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "APIResponse",
        "type": "object"
      },
...
      "CampaignAggregateResponse": {
        "description": "Per-campaign rollup (ports ``computeCampaignAggregate``, architecture \u00a77.3).\n\n``engagement = likes + comments``; ``estimated_reach`` is the org's follower\ncount (v1 reach approximation).",
        "properties": {
          "comments": {
            "title": "Comments",
            "type": "integer"
          },
          "engagement": {
            "title": "Engagement",
            "type": "integer"
          },
          "estimatedReach": {
            "title": "Estimatedreach",
            "type": "integer"
          },
          "likes": {
            "title": "Likes",
            "type": "integer"
          },
          "postCount": {
            "title": "Postcount",
            "type": "integer"
          }
        },
        "required": [
          "postCount",
          "likes",
          "comments",
          "engagement",
          "estimatedReach"
        ],
        "title": "CampaignAggregateResponse",
        "type": "object"
      },
      "CampaignDetailResponse": {
        "description": "``GET /api/campaigns/{id}`` \u2014 full application joined with the drop.",
        "properties": {
          "allocatedUnits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allocatedunits"
          },
          "appliedAt": {
            "title": "Appliedat",
            "type": "integer"
          },
          "applyCloseAt": {
            "title": "Applycloseat",
            "type": "integer"
          },
          "applyOpenAt": {
            "title": "Applyopenat",
            "type": "integer"
          },
          "brandName": {
            "title": "Brandname",
            "type": "string"
          },
          "brandTrackerStage": {
            "title": "Brandtrackerstage",
            "type": "string"
          },
          "capacityTotal": {
            "title": "Capacitytotal",
            "type": "integer"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "decisionAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisionat"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dropId": {
            "format": "uuid",
            "title": "Dropid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image": {
            "title": "Image",
            "type": "string"
          },
          "orgId": {
            "format": "uuid",
            "title": "Orgid",
            "type": "string"
          },
          "pitch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitch"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "totalProductUnits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalproductunits"
          },
          "trackingNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trackingnumber"
          }
        },
        "required": [
          "id",
          "dropId",
          "orgId",
          "decision",
          "pitch",
          "trackingNumber",
          "allocatedUnits",
          "appliedAt",
          "decisionAt",
          "title",
          "description",
          "brandName",
          "image",
          "brandTrackerStage",
          "applyOpenAt",
          "applyCloseAt",
          "capacityTotal",
          "totalProductUnits"
        ],
        "title": "CampaignDetailResponse",
        "type": "object"
      },
      "CampaignListItem": {
        "description": "One row of ``GET /api/campaigns`` (application + joined drop fields).",
        "properties": {
          "allocatedUnits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allocatedunits"
          },
          "appliedAt": {
            "title": "Appliedat",
            "type": "integer"
          },
          "brandName": {
            "title": "Brandname",
            "type": "string"
          },
          "brandTrackerStage": {
            "title": "Brandtrackerstage",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "decisionAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisionat"
          },
          "dropId": {
            "format": "uuid",
            "title": "Dropid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image": {
            "title": "Image",
            "type": "string"
          },
          "pitch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitch"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trackingNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trackingnumber"
          }
        },
        "required": [
          "id",
          "dropId",
          "decision",
          "pitch",
          "trackingNumber",
          "allocatedUnits",
          "appliedAt",
          "decisionAt",
          "title",
          "brandName",
          "brandTrackerStage",
          "image"
        ],
        "title": "CampaignListItem",
        "type": "object"
      },
...
      "DataResponse_CampaignAggregateResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignAggregateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[CampaignAggregateResponse]",
        "type": "object"
      },
      "DataResponse_CampaignDetailResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignDetailResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[CampaignDetailResponse]",
        "type": "object"
      },
...
      "DataResponse_OkResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OkResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[OkResponse]",
        "type": "object"
      },
...
      "DataResponse_PostResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[PostResponse]",
        "type": "object"
      },
...
      "DataResponse_list_CampaignListItem__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CampaignListItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[list[CampaignListItem]]",
        "type": "object"
      },
...
      "DataResponse_list_SuggestionResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SuggestionResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "title": "DataResponse[list[SuggestionResponse]]",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Machine-readable error payload returned in `APIResponse.error`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "LinkPostRequest": {
        "description": "Body for ``POST|DELETE /api/campaigns/{id}/link-post``.",
        "properties": {
          "postId": {
            "format": "uuid",
            "title": "Postid",
            "type": "string"
          }
        },
        "required": [
          "postId"
        ],
        "title": "LinkPostRequest",
        "type": "object"
      },
...
      "OkResponse": {
        "description": "Generic success ack (logout, notify, unlink, dismiss, password flows).",
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "SuggestionResponse": {
        "description": "A pending auto-link suggestion joined with its post (architecture \u00a77.4.1).",
        "properties": {
          "caption": {
            "title": "Caption",
            "type": "string"
          },
          "comments": {
            "title": "Comments",
            "type": "integer"
          },
          "likes": {
            "title": "Likes",
            "type": "integer"
          },
          "matchEvidence": {
            "title": "Matchevidence",
            "type": "string"
          },
          "matchReason": {
            "title": "Matchreason",
            "type": "string"
          },
          "postId": {
            "format": "uuid",
            "title": "Postid",
            "type": "string"
          },
          "postedAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postedat"
          },
          "thumbnailUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnailurl"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "postId",
          "url",
          "thumbnailUrl",
          "caption",
          "postedAt",
          "likes",
          "comments",
          "matchReason",
          "matchEvidence"
        ],
        "title": "SuggestionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Buzz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/campaigns": {
      "get": {
        "description": "The caller org's campaigns (non-denied applications + joined drop).",
        "operationId": "list_campaigns_api_campaigns_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_CampaignListItem__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Campaigns",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{application_id}": {
      "get": {
        "description": "One campaign owned by the caller org (404 otherwise).",
        "operationId": "campaign_detail_api_campaigns__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_CampaignDetailResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Campaign Detail",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{application_id}/aggregate": {
      "get": {
        "description": "Per-campaign metric rollup (postCount/likes/comments/engagement/reach).",
        "operationId": "campaign_aggregate_api_campaigns__application_id__aggregate_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_CampaignAggregateResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Campaign Aggregate",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{application_id}/link-post": {
      "delete": {
        "description": "Unlink a post from this campaign (idempotent).",
        "operationId": "unlink_campaign_post_api_campaigns__application_id__link_post_delete",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Unlink Campaign Post",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "description": "Manually link one of the caller's posts (409 ``POST_ALREADY_LINKED``).",
        "operationId": "link_campaign_post_api_campaigns__application_id__link_post_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PostResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Link Campaign Post",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{application_id}/suggestions": {
      "get": {
        "description": "Pending auto-link suggestions for this campaign.",
        "operationId": "campaign_suggestions_api_campaigns__application_id__suggestions_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_SuggestionResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Campaign Suggestions",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{application_id}/suggestions/{post_id}/accept": {
      "post": {
        "description": "Confirm a suggestion and link the post (409/404).",
        "operationId": "accept_campaign_suggestion_api_campaigns__application_id__suggestions__post_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Post Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PostResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Accept Campaign Suggestion",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{application_id}/suggestions/{post_id}/dismiss": {
      "post": {
        "description": "Reject a pending suggestion (404 ``SUGGESTION_NOT_FOUND``).",
        "operationId": "dismiss_campaign_suggestion_api_campaigns__application_id__suggestions__post_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Post Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OkResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Dismiss Campaign Suggestion",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  }
}